Fix Raw Block Serialization
Fixes an issue whereby a RawBlock is serialized in the RPC call '/getblocks' that results in the transactions property only containing ONE transaction in the block. This creates the transactions property as an array and loads the transaction information (blob + size) into properties of an object that is added to the array of transactions.